restructure files, add holiday support
