+++ .gitignore
@@ -1,13 +1,1 @@
-*.sw?
-.DS_Store
-coverage
-rdoc
-pkg
-*.bundle
-*.so
-*.dll
-*.o
-Makefile
-*.gemspec
-*.gem
-
+/Gemfile.lock
