+++ Rakefile
@@ -1,3 +1,5 @@
-require 'rubygems'
-require 'rake'
+require "bundler/gem_tasks"
+require 'rspec/core/rake_task'
+RSpec::Core::RakeTask.new :spec
 
+task default: :spec
