+++ lib/business_days/version.rb
@@ -1,0 +1,3 @@
+module BusinessDays
+  VERSION = "0.1.0"
+end
