+++ lib/ext/date.rb
@@ -1,0 +1,3 @@
+class Date
+  include BusinessDays::Calculations
+end
