+++ lib/ext/numeric.rb
@@ -1,0 +1,3 @@
+class Numeric
+  include BusinessDays
+end
