+++ lib/ext/time.rb
@@ -1,0 +1,3 @@
+class Time
+  include BusinessDays::Calculations
+end
