+++ spec/spec_helper.rb
@@ -1,11 +1,1 @@
-require 'rubygems'
-
-# require something.
-
-gem "rspec", ">=1.2.5"
-
-Spec::Runner.configure do |config|
-  
-end
-
-
+require "business_days"
